Langchain/requirements.txt
langchain
python-dotenv
Ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
langchain-community
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain
langchain_groq
langchain_core
fastapi
uvicorn
langserve[all]
sse_starlette